Move activity manager to the namespace
